create supplier edit component
